requirements.txt
django-randomslugfield
memcached
Pillow